requirements/dev.txt
-r base.txt
django-debug-toolbar==1.2.1
# Required by django-debug-toolbar
sqlparse==0.1.11
